Put json converters in more reasonable places, improve the WindowLayout ct converter.
